계좌 데이터 관련 프로퍼티 타입 Float에서 Double로 변경

+++ Virtual Investment/Virtual Investment/DataBase/AmountData.swift
@@ -9,9 +9,9 @@ import Foundation
 
 class AmountData {
   static var shared = AmountData()
-  var deposit: Float = 0
-  var evaluatedPrice: Float = 0
-  var investmentAccount: Float = 0
+  var deposit: Double = 0
+  var evaluatedPrice: Double = 0
+  var investmentAccount: Double = 0
   var investededCoins: [Coin] = []
 
   private init() {
